Move tests to corresponding folder structure
